examples/denylist/pyproject.toml
[project]
name = "denylist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "opadk",
]

[tool.uv.sources]
opadk = { workspace = true }
